wallet-ext: AccountAddress component (#8907)
updates AccountAddress component

* update styles
* remove unnecessary props
* remove CopyToClipboard component

no visible changes

part of APPS-510
