--- apps/wallet/src/ui/app/components/account-address/AccountAddress.module.scss
@@ -1,32 +1,0 @@
-@use '_values/colors';
-
-.address-container {
-    display: inline-flex;
-    flex-direction: row;
-    align-items: center;
-    max-width: 100%;
-}
-
-.address {
-    font-weight: 500;
-    overflow-wrap: break-word;
-    flex: 1;
-    min-width: 0;
-
-    &.normal {
-        color: colors.$sui-dark-blue;
-        font-size: 13px;
-    }
-
-    &.faded {
-        color: colors.$gray-70;
-        opacity: 1;
-        font-size: 13px;
-    }
-}
-
-.explorer-link {
-    color: colors.$gray-60;
-    margin-left: 5px;
-    font-size: 9px;
-}
