--- apps/wallet/src/ui/app/components/copy-to-clipboard/CopyToClipboard.module.scss
@@ -1,29 +1,0 @@
-@use '_values/colors';
-
-.container {
-    display: inline-flex;
-    flex-flow: row nowrap;
-    align-items: center;
-    max-width: 100%;
-    flex: 1;
-}
-
-.copy-icon {
-    margin-left: 6px;
-    cursor: pointer;
-    transition: all 0.3s ease-in-out;
-
-    &.normal {
-        color: colors.$gray-60;
-        font-size: 12px;
-    }
-
-    &.highlighted {
-        color: colors.$sui-steel-blue;
-        font-size: 12px;
-    }
-
-    &.plain {
-        font-weight: 100;
-    }
-}
